create a file for the application info page handlers , open the db, get all of the necessary elements
