+++ team/m3/macy-graves/feature-sequence-diagram.md
@@ -1,0 +1,9 @@
+The feature that took me the most time and energy is Application Information form. The idea for this feature is that when you have an application on your account, you can double click it to direct you to the Application Information Page. When double clicked, all the information for that specific application is listed on the new page.
+
+My goal was that after the add application popup was completed, the application info was saved in IndexedDB.
+
+
+Description
+
+double click on an application box on the home page.
+brings you to the application page with all of that applications information loaded into the page.
